MAINT: fix `RecursionError` in `Generator.throw()`
When `yf` is a `Generator`, caller of `throw` should be `yf`
instead of `self`. Otherwise, it may leads to the following error:
`RecursionError: maximum recursion depth exceeded in comparison`

Besides, `f_lasti` in our implementation is different from the one
in CPython implementation. Initial values in both of CPython and
ours are -1 and 0 respectively. And that might be the reason that
`f_lasti` might exceed the length of `f_code.co_code`.