MAINT: rewrite `MAKE_FUNCTION` for setting up additional function attributes
This commit is mainly providing patches for 2 unhandled function
attributes: `__annotations__`, `__kwdefaults__`.

Note that `__kwdefaults__` is a special attributes for setting up
keyword-only functions, e.g. ```def fn(*, a=1, b=2):```

And there will be a new change in Py38 that possibly affect
`Function.__call__` implementation. Check the comment in it for
more details.