BUG: `why` should be checked before dispatching next operation
While value of `why` is `extended_arg`, execution loop of bytecode
operation will start over from the head of the while-loop. But the
value of `why` should be check whether it is None or not. Otherwise,
it will be overwritten by the next `vm.dispatch()`, and it may makes
the exception raised from the previous operation being ignored.