BUG: `EXTENDED_ARG` does not change offset according to given argument
In previous version, the offset evaluated by `EXTENDED_ARG` is
always 2 bytes, which is valid only when `EXTENDED_ARG` is used for
extending the argument existing at the index of stack over 2**16.
However, this operation can also be used to process the function
annotations. Where the argument `count` depends on the amount of
annotations.