BUG: failed to raised a new exception from an outer `finally` block
In prior design of exception handling process, we didn't discriminate
whether an exception is newly created or is cached by vm. And any
exception is stored in to `GlobalCache()._content['last_exception']`.

That leads to the issue mentioned in the title of this commit.

To emulate the exception handling process in CPython, we use
`new_exception` and `last_exception` as the implementation
of `tstate->exc_[type/value/traceback]` and
`tstate->curexc_[type/value/traceback]` respectively.

Note that 'new_exception' denotes the current exception hold by vm,
and 'last_exception' denotes the exception should be raised.