UPDATE: implement a switch for debugging with/without entering virtual machine
In prior revisions, the invocation of `pdb.set_trace()` always
make user able to step into the execution flow of virtual machine.

In order to make it work like the real Python runtime, mechanism of
line tracing is implemented.

Besides, some tricks are made in `_c_api` module due to the
restrictions of Python. e.g. create a `frame` object...

For now, user can run this vm with `pdb` normally like using the
real Python runtime. And it also possible to explore the internal
exection of vm with `pdb` by setting a environment variable
`DEBUG_INTERNAL=1` before running this vm.